+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.2 (2016-11-11)
+
+### Api
+- Core done:
+    - Architecture splitted
+    - First entity created (rider)
+    - Database connection
+    - Controller action "getAll()" created just to test
+
 ## 0.0.1 (2016-11-07)
 
 ### App